dom/base/test/test_bug420700.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=420700
-->
<head>
  <title>Test for Bug 420700</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=420700">Mozilla Bug 420700</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

  var r = document.createRange();
  r.selectNode(document.documentElement);

  var df = r.createContextualFragment("<p>BAD</p>");

  var display = document.getElementById("display");
  display.innerHTML = "<p>GOOD</p>";

  var p = display.firstChild;

  is(p.textContent, "GOOD", "createContextualFragment tests");
</script>
</pre>
</body>
</html>
